Disconnect  featureAdded signal. Move code from plugin to new_plan.
Disconnect featureAdded signal. Fixes commit errors.
Moved code from plugin.py to new_plan.py
