+++ arho_feature_template/plugin.py
@@ -3,14 +3,13 @@ from __future__ import annotations
 import os
 from typing import TYPE_CHECKING, Callable, cast
 
-from qgis.core import QgsProject, QgsVectorLayer
 from qgis.PyQt.QtCore import QCoreApplication, Qt, QTranslator
 from qgis.PyQt.QtGui import QIcon
 from qgis.PyQt.QtWidgets import QAction, QWidget
 from qgis.utils import iface
 
 from arho_feature_template.core.feature_template_library import FeatureTemplater, TemplateGeometryDigitizeMapTool
-from arho_feature_template.core.update_plan import LandUsePlan, update_selected_plan
+from arho_feature_template.core.new_plan import NewPlan
 from arho_feature_template.qgis_plugin_tools.tools.custom_logging import setup_logger, teardown_logger
 from arho_feature_template.qgis_plugin_tools.tools.i18n import setup_translation
 from arho_feature_template.qgis_plugin_tools.tools.resources import plugin_name
@@ -125,6 +124,7 @@ class Plugin:
 
     def initGui(self) -> None:  # noqa N802
         self.templater = FeatureTemplater()
+        self.new_plan = NewPlan()
 
         plan_icon_path = os.path.join(PLUGIN_PATH, "resources/icons/city.png")  # A placeholder icon
         load_icon_path = os.path.join(PLUGIN_PATH, "resources/icons/folder.png")  # A placeholder icon
@@ -165,60 +165,8 @@ class Plugin:
         if not isinstance(new_tool, TemplateGeometryDigitizeMapTool):
             self.template_dock_action.setChecked(False)
 
-    def clear_all_filters(self):
-        """Clear filters for all vector layers in the project."""
-        layers = QgsProject.instance().mapLayers().values()
-
-        for layer in layers:
-            if isinstance(layer, QgsVectorLayer):
-                layer.setSubsetString("")
-
     def digitize_new_plan(self):
-        # Filtered layers are not editable, so clear filters first.
-        self.clear_all_filters()
-        # Find and set the "Kaava" layer
-        layers = QgsProject.instance().mapLayersByName("Kaava")
-        if not layers:
-            iface.messageBar().pushMessage("Error", "Layer 'Kaava' not found", level=3)
-            return
-
-        kaava_layer = layers[0]
-
-        if not kaava_layer.isEditable():
-            kaava_layer.startEditing()
-
-        iface.setActiveLayer(kaava_layer)
-
-        iface.actionAddFeature().trigger()
-        kaava_layer.featureAdded.connect(self.feature_added)
-
-    def feature_added(self):
-        kaava_layer = iface.activeLayer()
-        feature_ids_before_commit = kaava_layer.allFeatureIds()
-        if kaava_layer.isEditable():
-            if not kaava_layer.commitChanges():
-                iface.messageBar().pushMessage("Error", "Failed to commit changes to the layer.", level=3)
-                return
-        else:
-            iface.messageBar().pushMessage("Error", "Layer is not editable.", level=3)
-            return
-        feature_ids_after_commit = kaava_layer.allFeatureIds()
-
-        # Finds the feature id that was committed by comparing ids before commit to features after commit.
-        new_feature_id = next((fid for fid in feature_ids_after_commit if fid not in feature_ids_before_commit), None)
-        if new_feature_id is not None:
-            new_feature = kaava_layer.getFeature(new_feature_id)
-
-            if new_feature.isValid():
-                feature_id_value = new_feature["id"]  # UUID of the new feature
-                iface.messageBar().pushMessage("Info", f"Feature 'id' field value: {feature_id_value}", level=0)
-
-                # plan = LandUsePlan(feature_id_value)
-                update_selected_plan(LandUsePlan(feature_id_value))
-            else:
-                iface.messageBar().pushMessage("Error", "Invalid feature retrieved.", level=3)
-        else:
-            iface.messageBar().pushMessage("Error", "No new feature was added.", level=3)
+        self.new_plan.add_new_plan()
 
     def load_existing_land_use_plan(self) -> None:
         """Open existing land use plan."""
